Moved out unit tests to root dir.
