+++ tests/test_user_model.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import create_access_token
 
 from api.main import db
 from models.main import User
-from api.test.base import BaseTestCase
+from tests.base import BaseTestCase
 
 
 class TestUserModel(BaseTestCase):
